// typescript/src/server/config/types.d.ts
export interface FileQueueOptions {
    inbox: string;
    outbox: string;
    pollIntervalMs?: number;
    archive?: string | false;
}